refactor(example-data): Move Link Target Data
Link Target Data Examples moved to example-data module.

Note, that this contains some duplications with still
existing example-data.js (`createLink`). In the end,
we should enhance our richtext factory functions, to
support proper escaping as `createLink` does.
