+++ packages/ckeditor5-coremedia-example-data/src/data/LinkTargetData.ts
@@ -1,0 +1,333 @@
+import { lorem } from "../LoremIpsum";
+import { em, h1, richtext, richTextDocument } from "../RichText";
+import { ExampleData } from "../ExampleData";
+
+const SOME_TARGET = "somewhere";
+const EVIL_TARGET = `<iframe src="javascript:alert('Boo 👻')" width="1px" height="1px">`;
+const LINK_TEXT = "Link";
+const MDASH = "\u{2014}";
+const HELLIP = "\u{2026}";
+const UNSET = MDASH;
+const EXAMPLE_URL = "https://example.org/";
+
+const serializer = new XMLSerializer();
+
+const truncate = (str: string | null, maxLength: number): string | null => {
+  if (!!str && str.length > maxLength) {
+    return `${str.substring(0, maxLength)}${HELLIP}`;
+  }
+  return str;
+};
+
+/**
+ * Escapes the given string for display in HTML.
+ * @param str - string to escape
+ * @returns escaped string
+ */
+const escape = (str: string | null): string | null => {
+  if (!str) {
+    return str;
+  }
+
+  const el = richTextDocument.createElement("span");
+  el.textContent = str;
+  // noinspection InnerHTMLJS
+  return el.innerHTML;
+};
+
+const tableHeader = (...headers: string[]): string =>
+  `<tr class="tr--header">${headers.map((h) => `<td class="td--header">${h}</td>`).join("")}</tr>`;
+
+const createLinkTableHeading = () =>
+  tableHeader("xlink:show", "xlink:role", "target", "Active Button", "Editor Value", "Link", "Comment");
+
+const renderUiEditorValue = (uiEditorValue: string | null): string => {
+  if (uiEditorValue === "") {
+    return `${em("empty")}`;
+  }
+  return uiEditorValue ?? UNSET;
+};
+
+// TODO: Move to `RichText` and add this proper attribute escaping to any
+//   generator functions such as `nonEmptyElement`.
+const createLink = (show: string | null | undefined, role: string | null | undefined, href = EXAMPLE_URL): string => {
+  const a = richTextDocument.createElement("a");
+  a.textContent = LINK_TEXT;
+  a.setAttribute("xlink:href", href);
+  show && a.setAttribute("xlink:show", show);
+  role && a.setAttribute("xlink:role", role);
+  return serializer.serializeToString(a);
+};
+
+interface Scenario {
+  comment?: string;
+  show: string | null;
+  role: string | null;
+  target: string | null;
+  uiActiveButton: string;
+  uiEditorValue: string | null;
+}
+
+const createLinkTableRow = ({ comment, show, role, target, uiActiveButton, uiEditorValue }: Scenario) => {
+  const shorten = (str: string | null): string | null => escape(truncate(str, 15));
+  return `<tr>
+    <td>${shorten(show) || UNSET}</td>
+    <td>${shorten(role) || UNSET}</td>
+    <td>${shorten(target) || UNSET}</td>
+    <td>${uiActiveButton || UNSET}</td>
+    <td>${renderUiEditorValue(shorten(uiEditorValue))}</td>
+    <td>${createLink(show, role, EXAMPLE_URL)}</td>
+    <td>${comment || ""}</td>
+  </tr>`;
+};
+
+const createLinkScenario = (title: string, scenarios: Scenario[]) => {
+  const scenarioTitle = h1(title);
+  const scenarioHeader = createLinkTableHeading();
+  const scenarioRows = scenarios.map(createLinkTableRow).join("");
+  return `${scenarioTitle}<table>${scenarioHeader}${scenarioRows}</table>`;
+};
+
+const linkTargetExamples = () => {
+  const LONG_TARGET = lorem({ words: 100 });
+
+  /**
+   * The mapping we agreed upon for `xlink:show` to some target value.
+   * `other` is skipped here, as it is used for special meaning, which is,
+   * that the `xlink:show` is ignored but `xlink:role` will take over representing
+   * the `target` attribute.
+   */
+  const show = {
+    /**
+     * Open in new tab. Nothing to argue about.
+     */
+    new: "_blank",
+    /**
+     * Maybe either `_top` or `_self`. In CoreMedia CAE context we decided to
+     * map `replace` to `_self` as this is, what is documented for example
+     * at MDN.
+     */
+    replace: "_self",
+    /**
+     * Artificial mapping, we require, as there is no such `target` to represent
+     * embedding links.
+     */
+    embed: "_embed",
+    /**
+     * Artificial mapping, we require, as there is no such `target` to represent
+     * explicitly unspecified link behavior.
+     */
+    none: "_none",
+  };
+  const standardScenarios: Scenario[] = [
+    {
+      comment: `default for having no target set; if triggered, will change target value to _self`,
+      show: null,
+      role: null,
+      target: null,
+      uiActiveButton: "Open in Current Tab",
+      uiEditorValue: null,
+    },
+    {
+      show: "new",
+      role: null,
+      target: show.new,
+      uiActiveButton: "Open in New Tab",
+      uiEditorValue: null,
+    },
+    {
+      show: "replace",
+      role: null,
+      target: show.replace,
+      uiActiveButton: "Open in Current Tab",
+      uiEditorValue: null,
+    },
+    {
+      comment: "artificial reserved word for 'target'.",
+      show: "embed",
+      role: null,
+      target: show.embed,
+      uiActiveButton: "Show Embedded",
+      uiEditorValue: null,
+    },
+    {
+      comment: `artificial state, as a 'role' would have been expected; on ${em(
+        "Save"
+      )} the empty editor value will trigger the deletion of target attribute value`,
+      show: "other",
+      role: null,
+      target: "_other",
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: "",
+    },
+    {
+      comment: "artificial reserved word for 'target' to reflect this XLink-state",
+      show: "none",
+      role: null,
+      target: show.none,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: show.none,
+    },
+    {
+      comment: "Open in Frame; normal state for a named target.",
+      show: "other",
+      role: SOME_TARGET,
+      target: SOME_TARGET,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: SOME_TARGET,
+    },
+    {
+      comment: "Open in Frame; UI challenge with long target value",
+      show: "other",
+      role: LONG_TARGET,
+      target: LONG_TARGET,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: LONG_TARGET,
+    },
+    {
+      comment: "Open in Frame; UI cross-site-scripting challenge",
+      show: "other",
+      role: EVIL_TARGET,
+      target: EVIL_TARGET,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: EVIL_TARGET,
+    },
+  ];
+  const artificialRichTextScenarios: Scenario[] = [
+    {
+      comment: "artificial state, where a role misses an expected show attribute",
+      show: null,
+      role: SOME_TARGET,
+      target: `_role_${SOME_TARGET}`,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: `_role_${SOME_TARGET}`,
+    },
+    {
+      comment: "artificial state with unexpected role attribute",
+      show: "new",
+      role: SOME_TARGET,
+      target: `${show.new}_${SOME_TARGET}`,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: `${show.new}_${SOME_TARGET}`,
+    },
+    {
+      comment: "artificial state with unexpected role attribute",
+      show: "replace",
+      role: SOME_TARGET,
+      target: `${show.replace}_${SOME_TARGET}`,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: `${show.replace}_${SOME_TARGET}`,
+    },
+    {
+      comment: "artificial state with unexpected role attribute",
+      show: "embed",
+      role: SOME_TARGET,
+      target: `${show.embed}_${SOME_TARGET}`,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: `${show.embed}_${SOME_TARGET}`,
+    },
+    {
+      comment: "artificial state with unexpected role attribute",
+      show: "none",
+      role: SOME_TARGET,
+      target: `${show.none}_${SOME_TARGET}`,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: `${show.none}_${SOME_TARGET}`,
+    },
+  ];
+  const reservedTargetScenarios: Scenario[] = [
+    {
+      show: "replace",
+      role: null,
+      target: show.replace,
+      uiActiveButton: "Open in Current Tab",
+      uiEditorValue: null,
+    },
+    {
+      show: "new",
+      role: null,
+      target: show.new,
+      uiActiveButton: "Open in New Tab",
+      uiEditorValue: null,
+    },
+    {
+      comment: "artificial regarding xlink-attributes",
+      show: "other",
+      role: "_parent",
+      target: "_parent",
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: "_parent",
+    },
+    {
+      comment: "artificial regarding xlink-attributes",
+      show: "other",
+      role: "_top",
+      target: "_top",
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: "_top",
+    },
+  ];
+  const cornerCaseScenarios: Scenario[] = [
+    {
+      comment: "trying to misuse reserved word _role; handled as any custom target",
+      show: "other",
+      role: "_role",
+      target: "_role",
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: "_role",
+    },
+    {
+      comment: "trying to misuse reserved word _role; handled as any custom target",
+      show: "other",
+      role: "_role_",
+      target: "_role_",
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: "_role_",
+    },
+    {
+      comment: `trying to misuse artificial handling of ${show.new}_[role] with empty role; handled as any custom target`,
+      show: "other",
+      role: `${show.new}_`,
+      target: `${show.new}_`,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: `${show.new}_`,
+    },
+    {
+      comment: `trying to misuse artificial handling of ${show.replace}_[role] with empty role; handled as any custom target`,
+      show: "other",
+      role: `${show.replace}_`,
+      target: `${show.replace}_`,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: `${show.replace}_`,
+    },
+    {
+      comment: `trying to misuse artificial handling of ${show.embed}_[role] with empty role; handled as any custom target`,
+      show: "other",
+      role: `${show.embed}_`,
+      target: `${show.embed}_`,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: `${show.embed}_`,
+    },
+    {
+      comment: `trying to misuse artificial handling of ${show.none}_[role] with empty role; handled as any custom target`,
+      show: "other",
+      role: `${show.none}_`,
+      target: `${show.none}_`,
+      uiActiveButton: "Open in Frame",
+      uiEditorValue: `${show.none}_`,
+    },
+  ];
+  const scenarios = [
+    createLinkScenario("Standard Links", standardScenarios),
+    createLinkScenario("Artificial Richtext Scenarios", artificialRichTextScenarios),
+    createLinkScenario("Reserved Target Scenarios", reservedTargetScenarios),
+    createLinkScenario("Corner Case Scenarios", cornerCaseScenarios),
+  ].join("");
+  // noinspection XmlUnusedNamespaceDeclaration
+  return richtext(scenarios);
+};
+
+// noinspection JSUnusedGlobalSymbols Used in Example App
+export const linkTargetData: ExampleData = {
+  "Link Targets": linkTargetExamples(),
+};
